Replace ClonableSubtitleFile with pub enum
ClonableSubtitleFile was a private trait used as a workaround for trait
object limitations. However, it was exposed in the crate's public
interface, which made it hard to use in some cases.

Rather than expose the trait publicly (as it is mostly a workaround),
this replaces it with an enum `SubtitleFileVariant` which contains all
the SubtitleFile variants and implements the SubtitleFile trait.